superwand-0.1.2/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "superwand"
version = "v0.1.2"
description = "A tool for retheming images and CSS using color palettes and gradients."
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "Julian Henry"}
]
dependencies = [
    "numpy",
    "pillow",
    "scipy",
    "matplotlib",
    "scikit-learn",
    "flask",
    "werkzeug",
]

[project.optional-dependencies]
test = ["pytest", "pytest-cov"]

[tool.setuptools.package-data]
superwand = ["assets/fonts/*.ttf", "assets/themes_jpgs/*.jpg", "studio/templates/*.html", "studio/static/**/*"]

[project.scripts]
superwand = "superwand.core.superwand:main"
gradient-enforce = "superwand.utils.gradient_enforce:main"

[tool.setuptools.packages.find]
where = ["src"]
